// packages/mods/src/schema.ts
export type Tag = string;

export interface ModTypeProps {
  primary: number;
}

export interface SpawnWeightProps {
  value: number;
  tag: Tag;
}

export interface StatProps {
  id: string;
}

export interface ModProps {
  id: string;
  level: number;
  domain: number;
  name: string;
  correct_group: string;
  generation_type: number;
  stat1_min: number;
  stat1_max: number;
  stat2_min: number;
  stat2_max: number;
  stat3_min: number;
  stat3_max: number;
  stat4_min: number;
  stat4_max: number;
  stat5_min: number;
  stat5_max: number;
  mod_type: ModTypeProps;
  spawn_weights: SpawnWeightProps[];
  stats: StatProps[];
  tags: Tag[];
}
// ...
export type ItemClassId =
  | 'LifeFlask'
  | 'ManaFlask'
  | 'HybridFlask'
  | 'Currency'
  | 'Amulet'
  | 'Ring'
  | 'Claw'
  | 'Dagger'
  | 'Wand'
  | 'One Hand Sword'
  | 'Thrusting One Hand Sword'
  | 'One Hand Axe'
  | 'One Hand Mace'
  | 'Bow'
  | 'Staff'
  | 'Two Hand Sword'
  | 'Two Hand Axe'
  | 'Two Hand Mace'
  | 'Active Skill Gem'
  | 'Support Skill Gem'
  | 'Quiver'
  | 'Belt'
  | 'Gloves'
  | 'Boots'
  | 'Body Armour'
  | 'Helmet'
  | 'Shield'
  | 'SmallRelic'
  | 'MediumRelic'
  | 'LargeRelic'
  | 'StackableCurrency'
  | 'QuestItem'
  | 'Sceptre'
  | 'UtilityFlask'
  | 'UtilityFlaskCritical'
  | 'Map'
  | 'Unarmed'
  | 'FishingRod'
  | 'MapFragment'
  | 'HideoutDoodad'
  | 'Microtransaction'
  | 'Jewel'
  | 'DivinationCard'
  | 'LabyrinthItem'
  | 'LabyrinthTrinket'
  | 'LabyrinthMapItem'
  | 'MiscMapItem'
  | 'Leaguestone'
  | 'PantheonSoul'
  | 'UniqueFragment'
  | 'AbyssJewel'
  | 'IncursionItem';